+++ scripts/run.sh
@@ -8,6 +8,21 @@ config() {
   SERVICE="$(yq eval '.services.*.image | select(. == "sleede/fab-manager*") | path | .[-2]' docker-compose.yml)"
 }
 
+docker-compose()
+{
+  if ! docker compose version 1>/dev/null 2>/dev/null
+  then
+    if ! \docker-compose version 1>/dev/null 2>/dev/null
+    then
+      echo -e "\e[91m[ ❌ ] docker-compose was not found, exiting...\e[39m" && exit 1
+    else
+      \docker-compose "$@"
+    fi
+  else
+    docker compose "$@"
+  fi
+}
+
 run()
 {
   config