+++ setup/setup.sh
@@ -45,7 +45,7 @@ system_requirements()
       fi
     done
   fi
-  local _commands=("sudo" "curl" "sed" "openssl" "docker" "docker-compose" "systemctl")
+  local _commands=("sudo" "curl" "sed" "openssl" "docker" "systemctl")
   for _command in "${_commands[@]}"; do
     echo "detecting $_command..."
     if ! command -v "$_command"
@@ -54,7 +54,24 @@ system_requirements()
       echo -e "\e[91m[ ❌ ] $_command was not found, exiting...\e[39m" && exit 1
     fi
   done
+  echo "detecting docker-compose..."
+  docker-compose version
   printf "\e[92m[ ✔ ] All requirements successfully checked.\e[39m \n\n"
+}
+
+docker-compose()
+{
+  if ! docker compose version 1>/dev/null 2>/dev/null
+  then
+    if ! \docker-compose version 1>/dev/null 2>/dev/null
+    then
+      echo -e "\e[91m[ ❌ ] docker-compose was not found, exiting...\e[39m" && exit 1
+    else
+      \docker-compose "$@"
+    fi
+  else
+    docker compose "$@"
+  fi
 }
 
 is_root()