+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix a bug: Missing of description of PlanCategory migration
 - Fix a bug: Unable to create plans for all group
 - Fix a bug: backPrevLocation button on Notification Center opens reset password page (#348)
+- Fix a bug: compatibility with the new docker-compose CLI (#355)
 - Fix a security issue: updated moment.js to 2.29.2. to fix [CVE-2022-24785](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-24785)
 - Fix a security issue: updated nokogiri to 1.13.6 to fix [CVE-2022-298240](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-298240)
 - Fix a security issue: updated async to 2.6.4 to fix [CVE-2021-43138](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2021-43138)